+++ libbpf-cargo/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
   sub-command
 - Replaced `--debug` option of `libbpf` sub-command with `-v` /
   `--verbose`
+  - Removed `--quiet` option of `libbpf make` sub-command
 
 
 0.25.0-beta.1